Some package refactoring + adding missing javadoc/license
