+++ src/test/java/org/openhab/binding/jrule/internal/rules/timers/JRuleTimerTestRules.java
@@ -1,4 +1,16 @@
-package org.openhab.binding.jrule.internal.triggers;
+/**
+ * Copyright (c) 2010-2022 Contributors to the openHAB project
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ */
+package org.openhab.binding.jrule.internal.rules.timers;
 
 import org.openhab.automation.jrule.items.JRuleStringItem;
 import org.openhab.automation.jrule.rules.JRule;
@@ -6,7 +18,12 @@ import org.openhab.automation.jrule.rules.JRuleLogName;
 import org.openhab.automation.jrule.rules.JRuleName;
 import org.openhab.automation.jrule.rules.JRuleWhenItemChange;
 
-public class JRuleItemRules extends JRule {
+/**
+ * Rule used in testcase JRuleTimerTest
+ *
+ * @author Arne Seime - Initial contribution
+ */
+public class JRuleTimerTestRules extends JRule {
 
     public static final String TRIGGER_ITEM = "triggerItem";
     public static final String TARGET_ITEM = "targetItem";
